fits a calibration curve via gam

fits a calibration curve via gam

Usage

gam_cal(
  y,
  p,
  x,
  xp,
  time = NULL,
  save_data = TRUE,
  save_mod = TRUE,
  pw = FALSE,
  ...
)

Arguments

y

binary or a time-to-event (Surv) outcome. For the former family = binomial(link="logit") and for the latter family = mgcv::cox.ph().

p

predicted probabilities

x

predictor (could be transformation of p)

xp

values for plotting (same scale as x)

time

time to calculate survival probabilities at (only relevant if y is a Surv object)

save_data

whether to save the data elements in the returned object

save_mod

whether to save the model in the returned object

pw

save pointwise standard errors for plotting

...

additional arguments for mgcv::gam and mgcv::s

Value

list of class gam_cal

Examples

library(pmcalibration)
# simulate some data
n <- 500
dat <- sim_dat(N = n, a1 = .5, a3 = .2)
head(dat)
#>           x1         x2 y         LP
#> 1 -0.3872136  0.2360958 1  0.3305984
#> 2 -0.7854327  0.6289534 0  0.2447206
#> 3 -1.0567369  0.4179257 0 -0.2271387
#> 4 -0.7955414  1.9767585 0  1.3666984
#> 5 -1.7562754 -0.5062863 0 -1.5847261
#> 6 -0.6905379 -1.1099689 0 -1.1472116
# predictions
p <- with(dat, invlogit(.5 + x1 + x2 + x1*x2*.1))

gam_cal(y = dat$y, p = p, x = p, xp = NULL, k = 20, method="REML")
#> $y
#>   [1] 1 0 0 0 0 0 0 1 0 1 0 0 0 0 1 1 1 1 1 0 1 0 1 1 1 0 1 1 0 1 0 1 0 0 1 0 0
#>  [38] 0 1 1 1 0 0 1 1 1 0 1 0 1 1 1 1 0 1 0 0 0 1 1 0 1 1 0 1 1 1 0 1 1 0 0 1 0
#>  [75] 1 1 1 1 0 1 0 0 0 1 1 1 1 1 0 1 1 0 0 1 0 1 0 1 1 1 1 1 1 1 1 1 0 0 1 1 1
#> [112] 1 0 0 1 1 0 1 0 1 1 0 0 1 0 1 0 1 0 1 0 0 0 1 1 0 1 1 0 0 0 1 1 1 0 0 1 1
#> [149] 0 1 0 0 0 0 1 1 1 1 1 1 1 1 1 0 0 0 1 1 1 0 0 0 1 1 1 1 1 1 1 0 1 0 0 1 1
#> [186] 1 0 0 1 0 1 1 0 1 1 1 0 1 1 1 1 1 0 0 1 0 1 0 1 0 0 0 1 0 0 1 0 1 0 0 1 1
#> [223] 1 0 1 0 0 1 0 1 1 0 1 1 1 1 1 0 1 1 1 1 0 0 1 1 1 1 1 0 0 1 1 1 1 0 1 1 0
#> [260] 1 1 0 0 1 0 1 0 1 1 1 1 0 0 0 1 0 1 0 0 1 1 1 1 1 0 1 0 1 1 0 0 1 1 1 1 1
#> [297] 0 1 0 0 1 1 1 1 0 1 1 1 1 0 1 1 1 0 1 1 1 0 0 0 1 1 1 1 0 0 0 0 0 1 1 0 1
#> [334] 1 1 0 0 0 1 0 0 1 0 0 1 1 0 0 1 1 1 1 0 0 1 0 1 1 1 0 1 1 1 0 0 1 0 1 1 1
#> [371] 0 0 0 0 1 1 0 1 1 0 0 1 1 1 1 1 1 0 0 0 1 1 0 0 0 1 0 1 0 1 0 0 1 0 0 1 1
#> [408] 0 1 1 0 1 0 1 1 0 0 1 0 1 0 0 0 0 1 1 1 0 0 1 1 0 0 0 0 0 1 0 0 0 1 1 1 0
#> [445] 1 1 1 0 1 0 0 0 1 1 1 0 1 0 1 0 1 0 0 1 0 0 1 1 1 1 1 0 0 0 1 1 0 0 1 0 1
#> [482] 1 0 1 1 0 1 1 0 0 0 1 1 0 1 0 0 1 0 0
#> 
#> $p
#>   [1] 0.58412744 0.57300465 0.45438346 0.82112054 0.15793896 0.22725802
#>   [7] 0.27802921 0.60218796 0.47847940 0.85014129 0.28786029 0.04698994
#>  [13] 0.31664646 0.41220561 0.51810514 0.92269705 0.91036382 0.13013909
#>  [19] 0.80730037 0.49969896 0.98107273 0.37322733 0.95878480 0.83947494
#>  [25] 0.43387343 0.36086563 0.74351582 0.73616437 0.37724923 0.84002275
#>  [31] 0.30992423 0.53652003 0.40805882 0.36499320 0.58640914 0.43862068
#>  [37] 0.19852698 0.23452973 0.70194860 0.80435914 0.75324255 0.38173511
#>  [43] 0.95706588 0.60734760 0.92097492 0.86256295 0.39661179 0.56146829
#>  [49] 0.11434560 0.92427487 0.53641184 0.61254136 0.41919924 0.41801106
#>  [55] 0.84133794 0.87508375 0.48128399 0.28417471 0.86524031 0.89087302
#>  [61] 0.08391693 0.92776019 0.65514615 0.82840059 0.19820811 0.24136442
#>  [67] 0.27757263 0.73558002 0.64918422 0.97409107 0.14673627 0.47646961
#>  [73] 0.79701054 0.67958137 0.77113393 0.88038342 0.35973901 0.55678633
#>  [79] 0.49152582 0.66934955 0.13142153 0.07856512 0.14156644 0.86661180
#>  [85] 0.89242051 0.99389329 0.93293294 0.85598147 0.83424242 0.68624083
#>  [91] 0.67925627 0.35554383 0.57504879 0.42667684 0.61934036 0.51037385
#>  [97] 0.22748241 0.88974703 0.91369033 0.75808036 0.71614807 0.80731807
#> [103] 0.73705822 0.55554585 0.08785957 0.48058060 0.23784160 0.49166576
#> [109] 0.34759792 0.97621979 0.79533585 0.57218847 0.29528062 0.24276334
#> [115] 0.65548078 0.85876249 0.36156348 0.64387110 0.27498264 0.23613970
#> [121] 0.73512232 0.25976673 0.72864265 0.88195794 0.37366289 0.83350878
#> [127] 0.75289194 0.57854000 0.73348309 0.36473361 0.63330977 0.17968944
#> [133] 0.37376248 0.34661321 0.91196406 0.60228124 0.64235159 0.78668830
#> [139] 0.61086909 0.10553269 0.29221685 0.74012026 0.79826233 0.97316999
#> [145] 0.13978425 0.84623401 0.61223360 0.37189681 0.79835811 0.67796220
#> [151] 0.43251382 0.29387698 0.52500587 0.27101391 0.93827448 0.45357155
#> [157] 0.71548845 0.19161913 0.52131007 0.75727603 0.73434646 0.56590454
#> [163] 0.70308077 0.22362733 0.38856470 0.23216291 0.95466910 0.57826997
#> [169] 0.74549640 0.63954179 0.22542405 0.07664175 0.76896983 0.77319873
#> [175] 0.59522388 0.97028927 0.81607282 0.81501017 0.66368299 0.74903543
#> [181] 0.71503007 0.30366677 0.64550537 0.54646244 0.78992983 0.89724285
#> [187] 0.51239834 0.16481234 0.90633562 0.81251976 0.84277969 0.78837157
#> [193] 0.51470189 0.89262206 0.97536150 0.33014958 0.72033747 0.82172910
#> [199] 0.92997855 0.88257246 0.83536382 0.28163618 0.73115768 0.93293432
#> [205] 0.73875831 0.33257773 0.70764698 0.65113893 0.93473749 0.25683453
#> [211] 0.57410525 0.67940535 0.92883330 0.10956901 0.87533774 0.64107378
#> [217] 0.43282439 0.33135306 0.31839220 0.42667108 0.98232991 0.90755946
#> [223] 0.97176534 0.55081839 0.82817282 0.19655351 0.89771984 0.92212712
#> [229] 0.21245835 0.76689487 0.92376270 0.29246272 0.91156256 0.28972351
#> [235] 0.64130266 0.90413164 0.25265290 0.23623907 0.95470387 0.76384525
#> [241] 0.45898648 0.96149069 0.18124077 0.06918064 0.45496567 0.95750618
#> [247] 0.40475538 0.71781354 0.78888679 0.45819640 0.56333933 0.70622622
#> [253] 0.66799076 0.87512968 0.83403465 0.67235190 0.27612414 0.99147630
#> [259] 0.28682573 0.51323223 0.52775966 0.69201249 0.42633463 0.86297015
#> [265] 0.20810813 0.80845261 0.13309348 0.23654176 0.84314569 0.82387565
#> [271] 0.80377047 0.57526347 0.19469968 0.27619476 0.37061810 0.81043039
#> [277] 0.74587930 0.33982473 0.45911116 0.74143830 0.88639032 0.58311087
#> [283] 0.65654723 0.82991828 0.38865787 0.71027669 0.48756403 0.85338765
#> [289] 0.50557398 0.46077350 0.53438320 0.91821884 0.86677518 0.66420194
#> [295] 0.47995959 0.68075421 0.11561696 0.53925561 0.39535013 0.41815097
#> [301] 0.97929402 0.76641597 0.64615196 0.79469186 0.24098739 0.50578668
#> [307] 0.66557399 0.98318485 0.74087047 0.36330556 0.61367540 0.53313890
#> [313] 0.51589505 0.21923480 0.92399599 0.75873580 0.75207508 0.14266162
#> [319] 0.70448831 0.67725653 0.54273487 0.83966541 0.92730837 0.69241847
#> [325] 0.07588341 0.63576801 0.40106475 0.07806240 0.14137296 0.46056287
#> [331] 0.77202303 0.48000647 0.75513821 0.82313436 0.82464229 0.69979525
#> [337] 0.13939657 0.52656042 0.82173032 0.30666285 0.16042946 0.76263542
#> [343] 0.23237250 0.32156455 0.75386617 0.64641235 0.58968974 0.42338914
#> [349] 0.95683153 0.62078533 0.87589911 0.78332023 0.70998023 0.59263175
#> [355] 0.70444052 0.26403772 0.92877969 0.54172501 0.87219461 0.63294064
#> [361] 0.60460029 0.55216697 0.99142299 0.46739129 0.22921426 0.28543457
#> [367] 0.14821657 0.60715334 0.95174016 0.81762048 0.24643520 0.57435977
#> [373] 0.24730356 0.20965470 0.36477862 0.17383685 0.13086633 0.66978993
#> [379] 0.91897914 0.47061754 0.45830211 0.90012462 0.81893839 0.59255618
#> [385] 0.97632617 0.23958209 0.76999310 0.82152671 0.65156031 0.39259321
#> [391] 0.75596474 0.70263650 0.31044501 0.26682361 0.28890867 0.78850965
#> [397] 0.45858592 0.85244903 0.52405497 0.76260205 0.37185174 0.71749022
#> [403] 0.80695515 0.35172010 0.13620389 0.52121567 0.78547434 0.61113480
#> [409] 0.95840600 0.78697235 0.12508157 0.98624553 0.73183789 0.77806785
#> [415] 0.87151679 0.36323280 0.47509295 0.86297363 0.08325828 0.75381646
#> [421] 0.09934967 0.65499763 0.34774098 0.69911548 0.88236756 0.70013767
#> [427] 0.88273273 0.02939633 0.05106828 0.49032778 0.68835165 0.60487957
#> [433] 0.40758513 0.62745314 0.76823688 0.40991295 0.86795675 0.08718630
#> [439] 0.65788247 0.75124111 0.87890141 0.72833517 0.83670310 0.52316058
#> [445] 0.59666844 0.51792752 0.82637915 0.84666658 0.74285293 0.41711791
#> [451] 0.49041909 0.50622868 0.83334165 0.86105218 0.50860923 0.18564461
#> [457] 0.79358510 0.44033843 0.86348525 0.38004232 0.91456105 0.47774778
#> [463] 0.43365355 0.89555851 0.04108475 0.15491817 0.10034725 0.42799756
#> [469] 0.85442672 0.43145664 0.78167792 0.71021540 0.51876363 0.33963036
#> [475] 0.60206515 0.77084947 0.63785976 0.31992544 0.58017118 0.13279589
#> [481] 0.27484691 0.69444482 0.11480833 0.64231360 0.69361561 0.75243468
#> [487] 0.51595141 0.68915089 0.19947678 0.39634744 0.42068873 0.94437616
#> [493] 0.87565085 0.49213788 0.60868204 0.29608882 0.44640165 0.51638668
#> [499] 0.93050564 0.22895989
#> 
#> $x
#>   [1] 0.58412744 0.57300465 0.45438346 0.82112054 0.15793896 0.22725802
#>   [7] 0.27802921 0.60218796 0.47847940 0.85014129 0.28786029 0.04698994
#>  [13] 0.31664646 0.41220561 0.51810514 0.92269705 0.91036382 0.13013909
#>  [19] 0.80730037 0.49969896 0.98107273 0.37322733 0.95878480 0.83947494
#>  [25] 0.43387343 0.36086563 0.74351582 0.73616437 0.37724923 0.84002275
#>  [31] 0.30992423 0.53652003 0.40805882 0.36499320 0.58640914 0.43862068
#>  [37] 0.19852698 0.23452973 0.70194860 0.80435914 0.75324255 0.38173511
#>  [43] 0.95706588 0.60734760 0.92097492 0.86256295 0.39661179 0.56146829
#>  [49] 0.11434560 0.92427487 0.53641184 0.61254136 0.41919924 0.41801106
#>  [55] 0.84133794 0.87508375 0.48128399 0.28417471 0.86524031 0.89087302
#>  [61] 0.08391693 0.92776019 0.65514615 0.82840059 0.19820811 0.24136442
#>  [67] 0.27757263 0.73558002 0.64918422 0.97409107 0.14673627 0.47646961
#>  [73] 0.79701054 0.67958137 0.77113393 0.88038342 0.35973901 0.55678633
#>  [79] 0.49152582 0.66934955 0.13142153 0.07856512 0.14156644 0.86661180
#>  [85] 0.89242051 0.99389329 0.93293294 0.85598147 0.83424242 0.68624083
#>  [91] 0.67925627 0.35554383 0.57504879 0.42667684 0.61934036 0.51037385
#>  [97] 0.22748241 0.88974703 0.91369033 0.75808036 0.71614807 0.80731807
#> [103] 0.73705822 0.55554585 0.08785957 0.48058060 0.23784160 0.49166576
#> [109] 0.34759792 0.97621979 0.79533585 0.57218847 0.29528062 0.24276334
#> [115] 0.65548078 0.85876249 0.36156348 0.64387110 0.27498264 0.23613970
#> [121] 0.73512232 0.25976673 0.72864265 0.88195794 0.37366289 0.83350878
#> [127] 0.75289194 0.57854000 0.73348309 0.36473361 0.63330977 0.17968944
#> [133] 0.37376248 0.34661321 0.91196406 0.60228124 0.64235159 0.78668830
#> [139] 0.61086909 0.10553269 0.29221685 0.74012026 0.79826233 0.97316999
#> [145] 0.13978425 0.84623401 0.61223360 0.37189681 0.79835811 0.67796220
#> [151] 0.43251382 0.29387698 0.52500587 0.27101391 0.93827448 0.45357155
#> [157] 0.71548845 0.19161913 0.52131007 0.75727603 0.73434646 0.56590454
#> [163] 0.70308077 0.22362733 0.38856470 0.23216291 0.95466910 0.57826997
#> [169] 0.74549640 0.63954179 0.22542405 0.07664175 0.76896983 0.77319873
#> [175] 0.59522388 0.97028927 0.81607282 0.81501017 0.66368299 0.74903543
#> [181] 0.71503007 0.30366677 0.64550537 0.54646244 0.78992983 0.89724285
#> [187] 0.51239834 0.16481234 0.90633562 0.81251976 0.84277969 0.78837157
#> [193] 0.51470189 0.89262206 0.97536150 0.33014958 0.72033747 0.82172910
#> [199] 0.92997855 0.88257246 0.83536382 0.28163618 0.73115768 0.93293432
#> [205] 0.73875831 0.33257773 0.70764698 0.65113893 0.93473749 0.25683453
#> [211] 0.57410525 0.67940535 0.92883330 0.10956901 0.87533774 0.64107378
#> [217] 0.43282439 0.33135306 0.31839220 0.42667108 0.98232991 0.90755946
#> [223] 0.97176534 0.55081839 0.82817282 0.19655351 0.89771984 0.92212712
#> [229] 0.21245835 0.76689487 0.92376270 0.29246272 0.91156256 0.28972351
#> [235] 0.64130266 0.90413164 0.25265290 0.23623907 0.95470387 0.76384525
#> [241] 0.45898648 0.96149069 0.18124077 0.06918064 0.45496567 0.95750618
#> [247] 0.40475538 0.71781354 0.78888679 0.45819640 0.56333933 0.70622622
#> [253] 0.66799076 0.87512968 0.83403465 0.67235190 0.27612414 0.99147630
#> [259] 0.28682573 0.51323223 0.52775966 0.69201249 0.42633463 0.86297015
#> [265] 0.20810813 0.80845261 0.13309348 0.23654176 0.84314569 0.82387565
#> [271] 0.80377047 0.57526347 0.19469968 0.27619476 0.37061810 0.81043039
#> [277] 0.74587930 0.33982473 0.45911116 0.74143830 0.88639032 0.58311087
#> [283] 0.65654723 0.82991828 0.38865787 0.71027669 0.48756403 0.85338765
#> [289] 0.50557398 0.46077350 0.53438320 0.91821884 0.86677518 0.66420194
#> [295] 0.47995959 0.68075421 0.11561696 0.53925561 0.39535013 0.41815097
#> [301] 0.97929402 0.76641597 0.64615196 0.79469186 0.24098739 0.50578668
#> [307] 0.66557399 0.98318485 0.74087047 0.36330556 0.61367540 0.53313890
#> [313] 0.51589505 0.21923480 0.92399599 0.75873580 0.75207508 0.14266162
#> [319] 0.70448831 0.67725653 0.54273487 0.83966541 0.92730837 0.69241847
#> [325] 0.07588341 0.63576801 0.40106475 0.07806240 0.14137296 0.46056287
#> [331] 0.77202303 0.48000647 0.75513821 0.82313436 0.82464229 0.69979525
#> [337] 0.13939657 0.52656042 0.82173032 0.30666285 0.16042946 0.76263542
#> [343] 0.23237250 0.32156455 0.75386617 0.64641235 0.58968974 0.42338914
#> [349] 0.95683153 0.62078533 0.87589911 0.78332023 0.70998023 0.59263175
#> [355] 0.70444052 0.26403772 0.92877969 0.54172501 0.87219461 0.63294064
#> [361] 0.60460029 0.55216697 0.99142299 0.46739129 0.22921426 0.28543457
#> [367] 0.14821657 0.60715334 0.95174016 0.81762048 0.24643520 0.57435977
#> [373] 0.24730356 0.20965470 0.36477862 0.17383685 0.13086633 0.66978993
#> [379] 0.91897914 0.47061754 0.45830211 0.90012462 0.81893839 0.59255618
#> [385] 0.97632617 0.23958209 0.76999310 0.82152671 0.65156031 0.39259321
#> [391] 0.75596474 0.70263650 0.31044501 0.26682361 0.28890867 0.78850965
#> [397] 0.45858592 0.85244903 0.52405497 0.76260205 0.37185174 0.71749022
#> [403] 0.80695515 0.35172010 0.13620389 0.52121567 0.78547434 0.61113480
#> [409] 0.95840600 0.78697235 0.12508157 0.98624553 0.73183789 0.77806785
#> [415] 0.87151679 0.36323280 0.47509295 0.86297363 0.08325828 0.75381646
#> [421] 0.09934967 0.65499763 0.34774098 0.69911548 0.88236756 0.70013767
#> [427] 0.88273273 0.02939633 0.05106828 0.49032778 0.68835165 0.60487957
#> [433] 0.40758513 0.62745314 0.76823688 0.40991295 0.86795675 0.08718630
#> [439] 0.65788247 0.75124111 0.87890141 0.72833517 0.83670310 0.52316058
#> [445] 0.59666844 0.51792752 0.82637915 0.84666658 0.74285293 0.41711791
#> [451] 0.49041909 0.50622868 0.83334165 0.86105218 0.50860923 0.18564461
#> [457] 0.79358510 0.44033843 0.86348525 0.38004232 0.91456105 0.47774778
#> [463] 0.43365355 0.89555851 0.04108475 0.15491817 0.10034725 0.42799756
#> [469] 0.85442672 0.43145664 0.78167792 0.71021540 0.51876363 0.33963036
#> [475] 0.60206515 0.77084947 0.63785976 0.31992544 0.58017118 0.13279589
#> [481] 0.27484691 0.69444482 0.11480833 0.64231360 0.69361561 0.75243468
#> [487] 0.51595141 0.68915089 0.19947678 0.39634744 0.42068873 0.94437616
#> [493] 0.87565085 0.49213788 0.60868204 0.29608882 0.44640165 0.51638668
#> [499] 0.93050564 0.22895989
#> 
#> $xp
#> NULL
#> 
#> $p_c
#>   [1] 0.58015562 0.56471946 0.39837389 0.84126312 0.10983871 0.15452353
#>   [7] 0.19594617 0.60487759 0.43152419 0.86203512 0.20487307 0.06173726
#>  [13] 0.23272817 0.34268601 0.48726212 0.90412576 0.89788659 0.09535528
#>  [19] 0.83051104 0.46124647 0.92924321 0.29477868 0.92046432 0.85467941
#>  [25] 0.37086259 0.28042547 0.77336541 0.76597308 0.29953972 0.85506498
#>  [31] 0.22599461 0.51336021 0.33741094 0.28517034 0.58330391 0.37716329
#>  [37] 0.13442645 0.15998436 0.72940914 0.82814934 0.78289017 0.30490161
#>  [43] 0.91974743 0.61184891 0.90327549 0.87020453 0.32306102 0.54857642
#>  [49] 0.08790922 0.90489890 0.51320696 0.61882045 0.35167082 0.35013671
#>  [55] 0.85598721 0.87801899 0.43542904 0.20149153 0.87191053 0.88729272
#>  [61] 0.07502986 0.90658699 0.67396417 0.84670099 0.13421633 0.16525846
#>  [67] 0.19553882 0.76537837 0.66649091 0.92659414 0.10378343 0.42873116
#>  [73] 0.82213488 0.70365410 0.79964592 0.88120277 0.27913876 0.54199369
#>  [79] 0.44975368 0.69141151 0.09598405 0.07295196 0.10108985 0.87277701
#>  [85] 0.88816772 0.93387890 0.90904284 0.86592857 0.85095348 0.71146945
#>  [91] 0.70326945 0.27437946 0.56756663 0.36139502 0.62787292 0.47631786
#>  [97] 0.15468974 0.88665229 0.89960400 0.78751862 0.74500931 0.83052517
#> [103] 0.76688077 0.54024706 0.07659528 0.43444894 0.16252281 0.44995004
#> [109] 0.26550479 0.92741138 0.82074137 0.56358144 0.21180858 0.16635499
#> [115] 0.67438107 0.86774973 0.28122429 0.65975961 0.19324023 0.16121432
#> [121] 0.76491181 0.18015125 0.75823807 0.88213472 0.29529214 0.85042483
#> [127] 0.78255192 0.57241898 0.76323561 0.28487051 0.64618827 0.12248252
#> [133] 0.29540962 0.26441784 0.89871599 0.60500402 0.65782246 0.81340994
#> [139] 0.61658095 0.08398628 0.20892436 0.76997162 0.82317094 0.92623791
#> [145] 0.10017577 0.85937757 0.61840866 0.29321337 0.82325001 0.70173549
#> [151] 0.36906604 0.21048360 0.49704143 0.18975800 0.91151769 0.39727121
#> [157] 0.74429787 0.12993627 0.49180318 0.78675412 0.76411948 0.55479812
#> [163] 0.73067477 0.15185449 0.31316734 0.15819001 0.91873798 0.57204416
#> [169] 0.77532854 0.65422654 0.15317056 0.07221823 0.79767183 0.80151596
#> [175] 0.59540082 0.92511348 0.83740157 0.83657906 0.68451241 0.77880627
#> [181] 0.74380272 0.21985167 0.66183712 0.52743005 0.81618485 0.89085645
#> [187] 0.47918190 0.11370547 0.89577209 0.83463832 0.85699258 0.81485490
#> [193] 0.48244240 0.88828125 0.92708289 0.24667199 0.74949738 0.84172362
#> [199] 0.90764740 0.88249672 0.85175857 0.19918679 0.76084365 0.90904348
#> [205] 0.76860041 0.24923816 0.73574113 0.66895058 0.90988582 0.17770997
#> [211] 0.56625297 0.70344587 0.90710131 0.08576286 0.87817324 0.65618937
#> [217] 0.36947611 0.24794165 0.23449955 0.36138748 0.92971074 0.89641857
#> [223] 0.92569159 0.53358198 0.84653319 0.13313047 0.89111929 0.90384511
#> [229] 0.14388073 0.79576545 0.90464855 0.20915476 0.89850845 0.20659855
#> [235] 0.65648215 0.89459887 0.17427342 0.16129049 0.91875271 0.79293944
#> [241] 0.40464458 0.92158098 0.12343057 0.06943592 0.39916523 0.91993162
#> [247] 0.33323740 0.74679982 0.81529545 0.40356597 0.55120245 0.73417131
#> [253] 0.68976478 0.87804689 0.85080392 0.69503289 0.19425076 0.93302710
#> [259] 0.20391962 0.48036202 0.50094503 0.71814290 0.36094743 0.87046523
#> [265] 0.14087058 0.83142918 0.09680936 0.16152267 0.85724686 0.84333924
#> [271] 0.82767355 0.56786541 0.13192269 0.19431341 0.29171361 0.83299589
#> [277] 0.77570667 0.25700257 0.40481488 0.77129319 0.88472425 0.57875084
#> [283] 0.67570784 0.84781524 0.31328094 0.73863105 0.44420129 0.86421095
#> [289] 0.46953411 0.40708760 0.51033282 0.90190071 0.87287990 0.68514769
#> [295] 0.43358403 0.70503938 0.08848863 0.51723445 0.32149903 0.35031718
#> [301] 0.92857682 0.79532357 0.66265737 0.82020324 0.16496392 0.46983451
#> [307] 0.68682398 0.93002705 0.77072449 0.28322448 0.62033629 0.50856957
#> [313] 0.48413184 0.14867614 0.90476265 0.78814012 0.78176238 0.10165523
#> [319] 0.73224299 0.70089709 0.52215891 0.85481356 0.90636968 0.71860876
#> [325] 0.07193081 0.64936916 0.32860550 0.07275952 0.10099026 0.40679940
#> [331] 0.80045277 0.43364931 0.78471242 0.84278282 0.84391300 0.72699168
#> [337] 0.09997791 0.49924504 0.84172454 0.22277785 0.11122629 0.79181034
#> [343] 0.15834824 0.23774237 0.78349087 0.66298742 0.58781859 0.35710498
#> [349] 0.91964923 0.62978558 0.87851355 0.81049263 0.73830628 0.59185481
#> [355] 0.73218984 0.18375388 0.90707568 0.52072998 0.87625258 0.64570950
#> [361] 0.60814244 0.53548462 0.93300820 0.41617403 0.15597745 0.20264275
#> [367] 0.10456630 0.61158724 0.91748871 0.83859353 0.16926086 0.56660742
#> [373] 0.16995396 0.14193465 0.28492249 0.11896212 0.09571139 0.69194417
#> [379] 0.90228169 0.42062498 0.40371024 0.89243599 0.83960297 0.59175128
#> [385] 0.92745200 0.16386980 0.79860706 0.84157059 0.66947961 0.31809974
#> [391] 0.78550347 0.73017857 0.22651130 0.18613375 0.20584263 0.81497306
#> [397] 0.40409762 0.86358484 0.49569356 0.79177914 0.29316042 0.74645288
#> [403] 0.83023519 0.27008582 0.09836159 0.49166941 0.81236246 0.61693712
#> [409] 0.92030684 0.81365438 0.09291141 0.93114868 0.76154504 0.80587370
#> [415] 0.87583499 0.28314076 0.42682061 0.87046746 0.07477122 0.78344303
#> [421] 0.08132960 0.67377905 0.26566294 0.72622575 0.88237613 0.72737700
#> [427] 0.88259098 0.05620802 0.06308969 0.44807326 0.71392094 0.60851981
#> [433] 0.33681090 0.63855812 0.79699995 0.33976464 0.87362185 0.07632585
#> [439] 0.67736507 0.78095417 0.88031976 0.75791820 0.85271538 0.49442585
#> [445] 0.59737269 0.48701050 0.84520654 0.85967387 0.77270565 0.34898554
#> [451] 0.44820129 0.47045881 0.85030418 0.86923337 0.47382274 0.12615626
#> [457] 0.81927547 0.37945355 0.87079439 0.30287191 0.90004928 0.43050694
#> [463] 0.37057180 0.88992384 0.05982697 0.10817607 0.08175299 0.36312470
#> [469] 0.86490123 0.36767164 0.80905750 0.73856393 0.48819498 0.25679227
#> [475] 0.60471114 0.79938727 0.65206543 0.23606301 0.57468143 0.09666200
#> [481] 0.19312034 0.72092703 0.08811971 0.65777397 0.71997978 0.78211021
#> [487] 0.48421164 0.71484591 0.13505399 0.32273341 0.35359831 0.91427022
#> [493] 0.87836315 0.45061263 0.61364460 0.21257425 0.38758014 0.48482806
#> [499] 0.90789776 0.15578777
#> 
#> $metrics
#>       Eavg        E50        E90       Emax        ECI 
#> 0.03626367 0.02876786 0.07871263 0.08393371 0.19576061 
#> 
#> $p_c_plot
#> NULL
#> 
#> $p_c_plot_se
#> NULL
#> 
#> $model
#> 
#> Family: binomial 
#> Link function: logit 
#> 
#> Formula:
#> y ~ s(x, k = k, fx = fx, bs = bs)
#> 
#> Estimated degrees of freedom:
#> 1  total = 2 
#> 
#> REML score: 255.9903     
#> 
#> $smooth_args
#> $smooth_args$smooth
#> [1] "gam"
#> 
#> $smooth_args$bs
#> [1] "tp"
#> 
#> $smooth_args$k
#> [1] 20
#> 
#> $smooth_args$fx
#> [1] FALSE
#> 
#> $smooth_args$method
#> [1] "REML"
#> 
#> 
#> $time
#> NULL
#> 
#> $outcome
#> [1] "binary"
#> 
#> attr(,"class")
#> [1] "gam_cal"