calibration curve via loess

calibration curve via loess

Usage

loess_cal(p, y, x, xp, save_data = TRUE, save_mod = TRUE, pw = FALSE)

Arguments

p

predicted probabilities

y

binary outcome

x

predictor (could be transformation of p)

xp

values for plotting (same scale as x)

save_data

whether to save y, p, x, xp in the returned object

save_mod

whether to save the model in the returned object

pw

save pointwise standard errors for plotting

Value

list of class loess_cal

Examples

library(pmcalibration)
# simulate some data
n <- 500
dat <- sim_dat(N = n, a1 = .5, a3 = .2)

# predictions
p <- with(dat, invlogit(.5 + x1 + x2 + x1*x2*.1))

loess_cal(y = dat$y, p = p, x = p, xp = NULL)
#> $y
#>   [1] 1 1 0 1 1 0 0 1 0 0 1 1 1 0 0 1 1 0 1 1 1 1 0 0 1 0 0 0 1 0 1 1 0 1 0 0 1
#>  [38] 1 1 1 1 1 0 0 1 1 1 0 1 0 1 1 0 1 0 0 0 0 0 1 1 1 0 1 0 1 0 0 0 0 1 1 1 0
#>  [75] 1 1 1 1 1 1 1 0 0 1 0 0 1 0 1 1 0 0 0 1 1 1 1 1 0 1 1 1 0 0 1 0 0 1 0 0 0
#> [112] 0 1 1 0 0 1 0 1 0 1 1 0 1 1 0 1 1 0 0 1 1 1 1 0 1 1 1 1 0 0 1 1 1 1 0 1 1
#> [149] 0 1 0 1 0 1 0 0 0 1 0 0 0 1 1 1 0 1 1 1 1 1 1 1 1 1 1 1 0 1 1 0 1 1 0 1 1
#> [186] 1 0 1 0 1 1 0 1 1 0 0 1 0 1 1 1 0 1 1 0 1 1 0 0 1 1 1 1 0 1 1 0 1 1 1 1 0
#> [223] 1 0 1 0 1 1 1 1 1 1 1 1 1 1 1 0 0 1 0 0 1 0 1 1 1 0 0 1 1 0 0 1 0 1 0 1 1
#> [260] 1 0 1 0 1 1 1 1 1 0 0 0 1 1 0 0 1 1 1 1 0 0 0 0 0 1 1 0 0 1 1 0 0 0 1 0 0
#> [297] 0 1 1 1 0 1 1 1 1 0 1 1 1 1 1 1 1 1 1 0 1 1 0 0 0 1 1 1 1 0 0 1 0 1 1 1 0
#> [334] 1 1 0 0 1 0 1 1 1 0 0 1 0 0 1 0 0 0 0 1 0 1 1 1 1 1 1 1 1 1 0 1 0 0 1 1 0
#> [371] 0 0 1 1 1 1 1 1 1 1 1 0 0 1 0 1 1 1 1 1 1 1 1 0 0 1 1 0 1 1 1 1 0 1 0 1 1
#> [408] 0 1 0 0 1 0 0 1 0 1 0 1 1 1 0 0 1 1 1 1 0 0 1 0 0 1 0 1 1 1 0 0 1 0 1 1 1
#> [445] 1 1 1 0 1 0 0 0 0 1 1 1 1 1 0 0 0 0 1 1 1 0 1 0 1 1 1 1 1 1 1 1 1 1 0 1 0
#> [482] 1 0 1 0 1 0 1 1 0 1 0 1 1 0 0 1 0 1 0
#> 
#> $p
#>   [1] 0.92132997 0.67026370 0.17615418 0.97084433 0.58930444 0.60091819
#>   [7] 0.70319499 0.85698203 0.03781479 0.31428851 0.70931610 0.90332727
#>  [13] 0.89789573 0.39019322 0.15664410 0.81698086 0.82900079 0.56894346
#>  [19] 0.92120416 0.61686378 0.25006612 0.77326192 0.51614712 0.07238250
#>  [25] 0.36224909 0.27561383 0.71504080 0.37784998 0.86400715 0.41963536
#>  [31] 0.11952199 0.57627060 0.70980402 0.47975114 0.35127677 0.63334362
#>  [37] 0.82676649 0.82333985 0.90453368 0.83585618 0.48081964 0.23342963
#>  [43] 0.45068374 0.44592544 0.27665660 0.96312764 0.90378447 0.15429159
#>  [49] 0.82497191 0.13850154 0.57465151 0.86311361 0.63283793 0.57287967
#>  [55] 0.61809713 0.29652106 0.07961815 0.17872988 0.21109669 0.85859283
#>  [61] 0.89137109 0.68542246 0.10463733 0.89608692 0.16636902 0.61469616
#>  [67] 0.18458082 0.41468880 0.09335044 0.51187344 0.75291451 0.19158835
#>  [73] 0.75890005 0.80550325 0.16182822 0.96748417 0.38789328 0.63900522
#>  [79] 0.74979861 0.86748997 0.98056875 0.71512575 0.07859048 0.67326397
#>  [85] 0.25793168 0.61379431 0.95613845 0.67392463 0.64055610 0.34220117
#>  [91] 0.15908740 0.54330533 0.71684012 0.38684755 0.69465222 0.77297388
#>  [97] 0.71725451 0.98183038 0.37913944 0.85154330 0.11785159 0.91874664
#> [103] 0.46569803 0.21389746 0.88406757 0.50188351 0.27582432 0.84993153
#> [109] 0.18516798 0.38672502 0.45726713 0.20055512 0.59191418 0.83012383
#> [115] 0.54469125 0.44215417 0.81324569 0.54542807 0.86049359 0.46718920
#> [121] 0.84895888 0.74162419 0.50125285 0.67458837 0.11373657 0.21441122
#> [127] 0.91123397 0.48948812 0.66828735 0.73720020 0.52493637 0.79954571
#> [133] 0.91780960 0.85981985 0.43127341 0.85297599 0.88715541 0.85522480
#> [139] 0.73143944 0.83147079 0.39299029 0.59075876 0.67821714 0.87889834
#> [145] 0.77766702 0.58991399 0.56422065 0.42903779 0.14128043 0.81275516
#> [151] 0.58888890 0.86511576 0.25115496 0.89463909 0.43471261 0.11850416
#> [157] 0.31573380 0.91042242 0.27385790 0.08214272 0.37908747 0.92998145
#> [163] 0.66823731 0.87251865 0.69693437 0.86180043 0.71476116 0.73903346
#> [169] 0.69168588 0.83911924 0.64366328 0.78469183 0.48349187 0.72187598
#> [175] 0.53136838 0.95805367 0.38379718 0.64555166 0.73111819 0.40201555
#> [181] 0.64911164 0.77734001 0.06376600 0.75651206 0.94168865 0.99189940
#> [187] 0.36593096 0.63663468 0.60924196 0.68461924 0.61468485 0.24204430
#> [193] 0.39726338 0.75136804 0.44077378 0.79870529 0.82763490 0.29196733
#> [199] 0.88937900 0.81740812 0.80089226 0.36272330 0.88194858 0.90163033
#> [205] 0.25755703 0.75785820 0.81530951 0.35841269 0.44496241 0.64347599
#> [211] 0.79833372 0.87875903 0.94163475 0.65130727 0.69927823 0.98216578
#> [217] 0.61455144 0.72889668 0.75091622 0.86487968 0.73141595 0.35090528
#> [223] 0.92591220 0.48911586 0.65249531 0.16066816 0.70916986 0.64041982
#> [229] 0.58544763 0.77464527 0.92503948 0.75077728 0.65460105 0.46852590
#> [235] 0.71088153 0.84867772 0.93966053 0.35405820 0.86862286 0.88001294
#> [241] 0.08680445 0.08477307 0.98882356 0.28145548 0.18591190 0.61361793
#> [247] 0.56824610 0.91373398 0.68407584 0.94893069 0.68779622 0.33129949
#> [253] 0.79700604 0.96634980 0.42794355 0.94225415 0.49196380 0.70818035
#> [259] 0.74116031 0.75984028 0.08132085 0.86711277 0.18334460 0.62294901
#> [265] 0.75542629 0.44170165 0.91200580 0.44338099 0.45558438 0.76241704
#> [271] 0.35572626 0.69113453 0.78123082 0.43033224 0.26012468 0.62823273
#> [277] 0.75393284 0.94606372 0.83983358 0.22328580 0.42286614 0.14062985
#> [283] 0.27386468 0.11779016 0.80337194 0.51933611 0.36881090 0.70062996
#> [289] 0.83493151 0.99117144 0.57094728 0.06313638 0.80875914 0.94266032
#> [295] 0.12777784 0.47610234 0.69435159 0.79350717 0.95351438 0.91006601
#> [301] 0.85883962 0.74275562 0.69529663 0.47869818 0.64596593 0.45345700
#> [307] 0.95911570 0.95857285 0.82846807 0.85933010 0.59455924 0.82524196
#> [313] 0.78014102 0.91033691 0.91614756 0.82746074 0.95060273 0.80178441
#> [319] 0.80081621 0.72688469 0.67735460 0.97926357 0.85653477 0.44309686
#> [325] 0.27285135 0.63952102 0.59202533 0.22608738 0.92527633 0.56673735
#> [331] 0.54385298 0.46701425 0.05658487 0.55319107 0.99398514 0.68533460
#> [337] 0.54818728 0.94580220 0.55442897 0.97937863 0.80742145 0.79778397
#> [343] 0.69342993 0.41146598 0.72938135 0.82027706 0.41691595 0.88067461
#> [349] 0.25387985 0.61320603 0.05975770 0.42837669 0.52954638 0.22052528
#> [355] 0.94430254 0.87140169 0.84646390 0.75366560 0.91019260 0.53849062
#> [361] 0.66177238 0.37188042 0.79655486 0.15240485 0.96098201 0.03043969
#> [367] 0.65894887 0.74991890 0.70186533 0.41997344 0.71707470 0.55319242
#> [373] 0.90914641 0.52958748 0.73244619 0.97952602 0.88028276 0.68948481
#> [379] 0.79675247 0.45409922 0.31100461 0.09950736 0.40785965 0.83481471
#> [385] 0.75575167 0.83569698 0.76119345 0.59297679 0.77808017 0.81224094
#> [391] 0.93492160 0.65006009 0.95019220 0.63095629 0.11437315 0.50124412
#> [397] 0.17239869 0.31892591 0.80297566 0.95303787 0.61207155 0.84116687
#> [403] 0.46378385 0.87754530 0.73913123 0.46400531 0.50400174 0.80512167
#> [409] 0.31942962 0.24148657 0.54091649 0.78927540 0.23579890 0.22094229
#> [415] 0.98386149 0.17793230 0.85513175 0.46168215 0.69203616 0.88267026
#> [421] 0.86249310 0.34858501 0.18322245 0.88573263 0.68678062 0.96264287
#> [427] 0.33283612 0.83223973 0.71502666 0.44163432 0.28463186 0.68630496
#> [433] 0.59111921 0.31298746 0.72543282 0.08229440 0.25502266 0.08714235
#> [439] 0.32776552 0.88611011 0.21870166 0.80809301 0.87967981 0.55905150
#> [445] 0.61946102 0.78288941 0.98590395 0.21125546 0.80109439 0.29850169
#> [451] 0.45309806 0.28606683 0.32447386 0.72600260 0.45888179 0.32011740
#> [457] 0.41121898 0.81174941 0.11835578 0.52730461 0.47808793 0.59628405
#> [463] 0.73611274 0.77681207 0.93182988 0.55143480 0.22996905 0.19511506
#> [469] 0.79212844 0.93556131 0.68355777 0.82048125 0.31584176 0.54034343
#> [475] 0.59515152 0.90072967 0.79809964 0.39481008 0.34263428 0.70375377
#> [481] 0.38956986 0.81345003 0.38659075 0.37725564 0.46840088 0.76399859
#> [487] 0.63903049 0.73201965 0.85418058 0.34200847 0.91188564 0.59075801
#> [493] 0.88428820 0.67649442 0.36124889 0.40914519 0.76348505 0.55640937
#> [499] 0.43425633 0.59004395
#> 
#> $x
#>   [1] 0.92132997 0.67026370 0.17615418 0.97084433 0.58930444 0.60091819
#>   [7] 0.70319499 0.85698203 0.03781479 0.31428851 0.70931610 0.90332727
#>  [13] 0.89789573 0.39019322 0.15664410 0.81698086 0.82900079 0.56894346
#>  [19] 0.92120416 0.61686378 0.25006612 0.77326192 0.51614712 0.07238250
#>  [25] 0.36224909 0.27561383 0.71504080 0.37784998 0.86400715 0.41963536
#>  [31] 0.11952199 0.57627060 0.70980402 0.47975114 0.35127677 0.63334362
#>  [37] 0.82676649 0.82333985 0.90453368 0.83585618 0.48081964 0.23342963
#>  [43] 0.45068374 0.44592544 0.27665660 0.96312764 0.90378447 0.15429159
#>  [49] 0.82497191 0.13850154 0.57465151 0.86311361 0.63283793 0.57287967
#>  [55] 0.61809713 0.29652106 0.07961815 0.17872988 0.21109669 0.85859283
#>  [61] 0.89137109 0.68542246 0.10463733 0.89608692 0.16636902 0.61469616
#>  [67] 0.18458082 0.41468880 0.09335044 0.51187344 0.75291451 0.19158835
#>  [73] 0.75890005 0.80550325 0.16182822 0.96748417 0.38789328 0.63900522
#>  [79] 0.74979861 0.86748997 0.98056875 0.71512575 0.07859048 0.67326397
#>  [85] 0.25793168 0.61379431 0.95613845 0.67392463 0.64055610 0.34220117
#>  [91] 0.15908740 0.54330533 0.71684012 0.38684755 0.69465222 0.77297388
#>  [97] 0.71725451 0.98183038 0.37913944 0.85154330 0.11785159 0.91874664
#> [103] 0.46569803 0.21389746 0.88406757 0.50188351 0.27582432 0.84993153
#> [109] 0.18516798 0.38672502 0.45726713 0.20055512 0.59191418 0.83012383
#> [115] 0.54469125 0.44215417 0.81324569 0.54542807 0.86049359 0.46718920
#> [121] 0.84895888 0.74162419 0.50125285 0.67458837 0.11373657 0.21441122
#> [127] 0.91123397 0.48948812 0.66828735 0.73720020 0.52493637 0.79954571
#> [133] 0.91780960 0.85981985 0.43127341 0.85297599 0.88715541 0.85522480
#> [139] 0.73143944 0.83147079 0.39299029 0.59075876 0.67821714 0.87889834
#> [145] 0.77766702 0.58991399 0.56422065 0.42903779 0.14128043 0.81275516
#> [151] 0.58888890 0.86511576 0.25115496 0.89463909 0.43471261 0.11850416
#> [157] 0.31573380 0.91042242 0.27385790 0.08214272 0.37908747 0.92998145
#> [163] 0.66823731 0.87251865 0.69693437 0.86180043 0.71476116 0.73903346
#> [169] 0.69168588 0.83911924 0.64366328 0.78469183 0.48349187 0.72187598
#> [175] 0.53136838 0.95805367 0.38379718 0.64555166 0.73111819 0.40201555
#> [181] 0.64911164 0.77734001 0.06376600 0.75651206 0.94168865 0.99189940
#> [187] 0.36593096 0.63663468 0.60924196 0.68461924 0.61468485 0.24204430
#> [193] 0.39726338 0.75136804 0.44077378 0.79870529 0.82763490 0.29196733
#> [199] 0.88937900 0.81740812 0.80089226 0.36272330 0.88194858 0.90163033
#> [205] 0.25755703 0.75785820 0.81530951 0.35841269 0.44496241 0.64347599
#> [211] 0.79833372 0.87875903 0.94163475 0.65130727 0.69927823 0.98216578
#> [217] 0.61455144 0.72889668 0.75091622 0.86487968 0.73141595 0.35090528
#> [223] 0.92591220 0.48911586 0.65249531 0.16066816 0.70916986 0.64041982
#> [229] 0.58544763 0.77464527 0.92503948 0.75077728 0.65460105 0.46852590
#> [235] 0.71088153 0.84867772 0.93966053 0.35405820 0.86862286 0.88001294
#> [241] 0.08680445 0.08477307 0.98882356 0.28145548 0.18591190 0.61361793
#> [247] 0.56824610 0.91373398 0.68407584 0.94893069 0.68779622 0.33129949
#> [253] 0.79700604 0.96634980 0.42794355 0.94225415 0.49196380 0.70818035
#> [259] 0.74116031 0.75984028 0.08132085 0.86711277 0.18334460 0.62294901
#> [265] 0.75542629 0.44170165 0.91200580 0.44338099 0.45558438 0.76241704
#> [271] 0.35572626 0.69113453 0.78123082 0.43033224 0.26012468 0.62823273
#> [277] 0.75393284 0.94606372 0.83983358 0.22328580 0.42286614 0.14062985
#> [283] 0.27386468 0.11779016 0.80337194 0.51933611 0.36881090 0.70062996
#> [289] 0.83493151 0.99117144 0.57094728 0.06313638 0.80875914 0.94266032
#> [295] 0.12777784 0.47610234 0.69435159 0.79350717 0.95351438 0.91006601
#> [301] 0.85883962 0.74275562 0.69529663 0.47869818 0.64596593 0.45345700
#> [307] 0.95911570 0.95857285 0.82846807 0.85933010 0.59455924 0.82524196
#> [313] 0.78014102 0.91033691 0.91614756 0.82746074 0.95060273 0.80178441
#> [319] 0.80081621 0.72688469 0.67735460 0.97926357 0.85653477 0.44309686
#> [325] 0.27285135 0.63952102 0.59202533 0.22608738 0.92527633 0.56673735
#> [331] 0.54385298 0.46701425 0.05658487 0.55319107 0.99398514 0.68533460
#> [337] 0.54818728 0.94580220 0.55442897 0.97937863 0.80742145 0.79778397
#> [343] 0.69342993 0.41146598 0.72938135 0.82027706 0.41691595 0.88067461
#> [349] 0.25387985 0.61320603 0.05975770 0.42837669 0.52954638 0.22052528
#> [355] 0.94430254 0.87140169 0.84646390 0.75366560 0.91019260 0.53849062
#> [361] 0.66177238 0.37188042 0.79655486 0.15240485 0.96098201 0.03043969
#> [367] 0.65894887 0.74991890 0.70186533 0.41997344 0.71707470 0.55319242
#> [373] 0.90914641 0.52958748 0.73244619 0.97952602 0.88028276 0.68948481
#> [379] 0.79675247 0.45409922 0.31100461 0.09950736 0.40785965 0.83481471
#> [385] 0.75575167 0.83569698 0.76119345 0.59297679 0.77808017 0.81224094
#> [391] 0.93492160 0.65006009 0.95019220 0.63095629 0.11437315 0.50124412
#> [397] 0.17239869 0.31892591 0.80297566 0.95303787 0.61207155 0.84116687
#> [403] 0.46378385 0.87754530 0.73913123 0.46400531 0.50400174 0.80512167
#> [409] 0.31942962 0.24148657 0.54091649 0.78927540 0.23579890 0.22094229
#> [415] 0.98386149 0.17793230 0.85513175 0.46168215 0.69203616 0.88267026
#> [421] 0.86249310 0.34858501 0.18322245 0.88573263 0.68678062 0.96264287
#> [427] 0.33283612 0.83223973 0.71502666 0.44163432 0.28463186 0.68630496
#> [433] 0.59111921 0.31298746 0.72543282 0.08229440 0.25502266 0.08714235
#> [439] 0.32776552 0.88611011 0.21870166 0.80809301 0.87967981 0.55905150
#> [445] 0.61946102 0.78288941 0.98590395 0.21125546 0.80109439 0.29850169
#> [451] 0.45309806 0.28606683 0.32447386 0.72600260 0.45888179 0.32011740
#> [457] 0.41121898 0.81174941 0.11835578 0.52730461 0.47808793 0.59628405
#> [463] 0.73611274 0.77681207 0.93182988 0.55143480 0.22996905 0.19511506
#> [469] 0.79212844 0.93556131 0.68355777 0.82048125 0.31584176 0.54034343
#> [475] 0.59515152 0.90072967 0.79809964 0.39481008 0.34263428 0.70375377
#> [481] 0.38956986 0.81345003 0.38659075 0.37725564 0.46840088 0.76399859
#> [487] 0.63903049 0.73201965 0.85418058 0.34200847 0.91188564 0.59075801
#> [493] 0.88428820 0.67649442 0.36124889 0.40914519 0.76348505 0.55640937
#> [499] 0.43425633 0.59004395
#> 
#> $xp
#> NULL
#> 
#> $p_c
#>   [1] 0.9626989 0.6699426 0.1766650 1.0030631 0.5601595 0.5763390 0.7149617
#>   [8] 0.9005397 0.1273165 0.2596665 0.7232847 0.9464742 0.9414004 0.3194193
#>  [15] 0.1675775 0.8572764 0.8705272 0.5320250 0.9625886 0.5984808 0.2173228
#>  [22] 0.8048962 0.4621777 0.1363331 0.2959386 0.2334611 0.7310019 0.3088134
#>  [29] 0.9078501 0.3465424 0.1522036 0.5420901 0.7239453 0.4152211 0.2872186
#>  [36] 0.6210211 0.8680808 0.8643176 0.9475897 0.8779935 0.4165646 0.2073168
#>  [43] 0.3800494 0.3746004 0.2341378 0.9971489 0.9468975 0.1665285 0.8661116
#>  [50] 0.1597489 0.5398588 0.9069261 0.6203374 0.5374214 0.6001833 0.2473323
#>  [57] 0.1385020 0.1779164 0.1946630 0.9022251 0.9351902 0.6906014 0.1467491
#>  [64] 0.9396917 0.1720208 0.5954840 0.1808036 0.3417058 0.1428853 0.4565788
#>  [71] 0.7794329 0.1843430 0.7869215 0.8441191 0.1699247 1.0005037 0.3174139
#>  [78] 0.6286324 0.7755449 0.9114343 1.0103359 0.7311158 0.1381880 0.6740033
#>  [85] 0.2222055 0.5942356 0.9916788 0.6748999 0.6307027 0.2802039 0.1686777
#>  [92] 0.4977533 0.7334108 0.3165066 0.7032679 0.8045368 0.7339644 1.0112653
#>  [99] 0.3099036 0.8948110 0.1515712 0.9604252 0.3978535 0.1962009 0.9280883
#> [106] 0.4435482 0.2335976 0.8931025 0.1810967 0.3164005 0.3877482 0.1890013
#> [113] 0.5637927 0.8717545 0.4995574 0.3703553 0.8530580 0.5005189 0.9042069
#> [120] 0.3996671 0.8920691 0.7654024 0.4427293 0.6758014 0.1500351 0.1964845
#> [127] 0.9537103 0.4275648 0.6672784 0.7598901 0.4737128 0.8370827 0.9595960
#> [134] 0.9035053 0.3585025 0.8963256 0.9311097 0.8986951 0.7525677 0.8732245
#> [141] 0.3218764 0.5621838 0.6807426 0.9229714 0.8103811 0.5610078 0.5255883
#> [148] 0.3561447 0.1609090 0.8524993 0.5595813 0.9089939 0.2179944 0.9383169
#> [155] 0.3621827 0.1518176 0.2606939 0.9529756 0.2323249 0.1392816 0.3098596
#> [162] 0.9701852 0.6672110 0.9165594 0.7063975 0.9055651 0.7306268 0.7621867
#> [169] 0.6991968 0.8815243 0.6348299 0.8190701 0.4199372 0.7401028 0.4821526
#> [176] 0.9931887 0.3138758 0.6373240 0.7521549 0.3299447 0.6419953 0.8099748
#> [183] 0.1338771 0.7839315 0.9800110 1.0185691 0.2989256 0.6254554 0.5879214
#> [190] 0.6895002 0.5954683 0.2124364 0.3256695 0.7775021 0.3688184 0.8360803
#> [197] 0.8690324 0.2442547 0.9332684 0.8577547 0.8386840 0.2963216 0.9259995
#> [204] 0.9448982 0.2219713 0.7856167 0.8553969 0.2928592 0.3735100 0.6345819
#> [211] 0.8356363 0.9228325 0.9799665 0.6448551 0.7096082 1.0115118 0.5952837
#> [218] 0.7492885 0.7769384 0.9087506 0.7525375 0.2869280 0.9666884 0.4270891
#> [225] 0.6463956 0.1693952 0.7230867 0.6305210 0.5547970 0.8066210 0.9659328
#> [232] 0.7767651 0.6491304 0.4012992 0.7254023 0.8917701 0.9783333 0.2894039
#> [239] 0.9125941 0.9240808 0.1407523 0.1401065 1.0163590 0.2372718 0.1814691
#> [246] 0.5939913 0.5310718 0.9559619 0.6887554 0.9859204 0.6938578 0.2720028
#> [253] 0.8340465 0.9996342 0.3550009 0.9804770 0.4307360 0.7217453 0.7648293
#> [260] 0.7880992 0.1390265 0.9110474 0.1801884 0.6068596 0.7825729 0.3698504
#> [267] 0.9544073 0.3717289 0.3857635 0.7913277 0.2907225 0.6984400 0.8147995
#> [274] 0.3575065 0.2235799 0.6140852 0.7807054 0.9835961 0.8822951 0.2014586
#> [281] 0.3497830 0.1606362 0.2323293 0.1515480 0.8416164 0.4663612 0.3012839
#> [288] 0.7114577 0.8769901 1.0180477 0.5347689 0.1337031 0.8479084 0.9808112
#> [295] 0.1554045 0.4106559 0.7028554 0.8298314 0.9895964 0.9526524 0.9024828
#> [302] 0.7668001 0.7041519 0.4139000 0.6378697 0.3832707 0.9940224 0.9935966
#> [309] 0.8699445 0.9029947 0.5674775 0.8664082 0.8134505 0.9528981 0.9581195
#> [316] 0.8688416 0.9872671 0.8397415 0.8385937 0.7466750 0.6795663 1.0093711
#> [323] 0.9000708 0.3714101 0.2316756 0.6293217 0.5639475 0.2030581 0.9661381
#> [330] 0.5290128 0.4984659 0.3994540 0.1319359 0.5107549 1.0200576 0.6904809
#> [337] 0.5041354 0.9833831 0.5124040 1.0094563 0.8463566 0.8349787 0.7015907
#> [344] 0.3386390 0.7499156 0.8609432 0.3438643 0.9247378 0.2196813 0.5934207
#> [351] 0.1327817 0.3554528 0.4797631 0.1998962 0.9821585 0.9154263 0.8894107
#> [358] 0.7803714 0.9527672 0.4914752 0.6585666 0.3038189 0.8335050 0.1656945
#> [365] 0.9954813 0.1256820 0.6548301 0.7756949 0.7131463 0.3468785 0.7337242
#> [372] 0.5107567 0.9518168 0.4798170 0.7538584 1.0095654 0.9243489 0.6961754
#> [379] 0.8337422 0.3840213 0.2573461 0.1449638 0.3352857 0.8768633 0.7829800
#> [386] 0.8778208 0.7897946 0.5652728 0.8108943 0.8519124 0.9743734 0.6432327
#> [393] 0.9869370 0.6177884 0.1502707 0.4427180 0.1748620 0.2629762 0.8411492
#> [400] 0.9892165 0.5918484 0.8837317 0.3955365 0.9216202 0.7623087 0.3958039
#> [407] 0.4463025 0.8436722 0.2633381 0.2121010 0.4946414 0.8246885 0.2087117
#> [414] 0.2001313 1.0127547 0.1775276 0.8985973 0.3930075 0.6996777 0.9267123
#> [421] 0.9062835 0.2851198 0.1801278 0.9297209 0.6924642 0.9967731 0.2731444
#> [428] 0.8740626 0.7309829 0.3697754 0.2393645 0.6918117 0.5626857 0.2587449
#> [435] 0.7447793 0.1393288 0.2203913 0.1408604 0.2693950 0.9300899 0.1988716
#> [442] 0.8471366 0.9237496 0.5185986 0.6020636 0.8168488 1.0142441 0.1947498
#> [449] 0.8389238 0.2486812 0.3828520 0.2403147 0.2669876 0.7455242 0.3896630
#> [456] 0.2638329 0.3384067 0.8513503 0.1517615 0.4768215 0.4131357 0.5698811
#> [463] 0.7585198 0.8093184 0.9717596 0.5084230 0.2052976 0.1861579 0.8281609
#> [470] 0.9749112 0.6880456 0.8611685 0.2607708 0.4938940 0.5683029 0.9440583
#> [477] 0.8353564 0.3234859 0.2805347 0.7157239 0.3188744 0.8532905 0.3162842
#> [484] 0.3083123 0.4011463 0.7933092 0.6286662 0.7533121 0.8975961 0.2800568
#> [491] 0.9542989 0.5621828 0.9283051 0.6783943 0.2951326 0.3364731 0.7926658
#> [498] 0.5150510 0.3616908 0.5611887
#> 
#> $metrics
#>       Eavg        E50        E90       Emax        ECI 
#> 0.03791931 0.03865253 0.06905480 0.09524226 0.18294715 
#> 
#> $p_c_plot
#> NULL
#> 
#> $p_c_plot_se
#> NULL
#> 
#> $model
#> Call:
#> loess(formula = y ~ x)
#> 
#> Number of Observations: 500 
#> Equivalent Number of Parameters: 4.47 
#> Residual Standard Error: 0.4009 
#> 
#> $smooth_args
#> $smooth_args$smooth
#> [1] "loess"
#> 
#> 
#> attr(,"class")
#> [1] "loess_cal"